// src/types/route.ts
export enum ROUTE_TYPE {
  REGISTER = 'register',
  LOGIN = 'login',
  FORGOT = 'forgot',
  ACCOUNT = 'account',
  NEW_EMPLOYEE = 'new-employee',
}

export enum ROUTE_NAME {
  HOME = 'Главная',
  ERROR = 'Ошибка',
  REGISTER = 'Регистрация',
  LOGIN = 'Войти',
  FORGOT = 'Забыли пароль?',
  ACCOUNT = 'Профиль',
  NEW_EMPLOYEE = 'Новый сотрудник',
}

export enum ROUTE_ICON {
  HOME = 'home',
  REGISTER = 'person_add',
  LOGIN = 'login',
  FORGOT = 'lock_reset',
  ACCOUNT = 'account_circle',
  NEW_EMPLOYEE = 'badge',
}